requirements.txt
Flask
Werkzeug
gunicorn
requests
PyYAML
